requirements.txt
streamlit>=1.35.0
plotly>=5.20.0
pandas>=2.0.0
fpdf2
